Fix persistent scaffold configuration during tests
When executing the command automatically, several times, during
tests, the same configuration was applied and some tests failed. This
is because of the Configuration Loader package! Its service provider
binds the default parser factory as a singleton instance, meaning
that the same configuration is used, regardless of how many files are
parsed. Normally this isn't a problem, yet this became an issue
during tests, in this package, because the incorrect configuration
was used after the first execution of the command.

Therefore, the way of how the configuration is read has been changed.
Now, the file name is used to ensure that the correct entry is
loaded. The alternative was to empty the configuration inside the
loader, after the command executed all of its tasks. But this is a
bit to dangerous, since we do not know the usage context and might
manipulate data in an undesired way.