Add Template Data processor / handler task
The task is at this point not completed, there are many things that
still need to get done, before this is in order. The entire
properties processing should be moved to a specialised handler, which
can be used when asking for the destination of the files to be
generated (based on templates).