Allow handler to receive parameters
Due to the "special" property handler, the IoC needed to support
parameters, when attempting to resolve instances. This should had
been supported from the start, but I didn't needed then, so it got
taken for granted.

This BaseTask's resolveHandler has also been modified and now also
supports parameters, which is given further to the handler that is
being resolved.