Add custom style component
This custom style is for testing only. It allows settings and getting the
internal question helper used for asking questions. Thus, I can now
manipulate the input stream, during tests.

It is basically a nasty hack, but there does not seem to be a better way
of doing this, for the current version of Symfony Console.