Add utility for setting input stream
Basically, this is the only way to ensure that we can perform automatic
tests of interactive commands. Perhaps not the best way, yet
the SymfonyStyle component does make it very hard to specify a stream
directly.

This will have to change in the future, because Symfony has started to see
that the component is a bit to tricky to test, in terms of interaction.