Change service providers registration
Unfortunately the list of components that needs to be registered in the
IoC has grown a lot and there is more to come. The ScaffoldServiceProvider
had become too large. Therefore, it now extends the
AggregateServiceProvider and contains a list of all the required
providers.

The IoC now only needs to register a single "core" service provider
instead. This should be much easier to maintain and developers can make
use of the small individual providers instead, should they feel the need
to extend this application.